Title: Finished pagination
Summary: Finally added the pagination code, it works perfectly only need
to change the design but that will be the last thing to do, to the whole
site.

Gonna start creating the right_side_bar which will appear on almost
every page, containing the search bar and other things I'll think of.