Title: Fixed simple bug
Summary: Fixed a simple bug in the index file, if the user doesn't
follow any sub-forum then he will get a warning letting him know and it
will be shown to him the lastest posts. If the user follows one or more
sub-forums but those sub-forums don't have any posts, the user is
advised of it and is shown the lastest posts.